Refactor API keys to fit architecture guidelines
